+++ src/main/java/com/sofodev/armorplus/blocks/castle/base/BlockStoneBrick.java
@@ -6,7 +6,6 @@ package com.sofodev.armorplus.blocks.castle.base;
 
 import com.sofodev.armorplus.blocks.BlockProperties;
 import com.sofodev.armorplus.blocks.base.BlockBase;
-import com.sofodev.armorplus.blocks.base.ToolType;
 import com.sofodev.armorplus.blocks.castle.BrickColor;
 import com.sofodev.armorplus.iface.IModdedBlock;
 import net.minecraft.block.material.MapColor;
@@ -16,6 +15,8 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.IBlockAccess;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
+
+import static com.sofodev.armorplus.config.ModConfig.RegistryConfig.blocks;
 
 
 /**
@@ -28,7 +29,7 @@ public class BlockStoneBrick extends BlockBase implements IModdedBlock {
     public String name;
 
     public BlockStoneBrick(BrickColor brickColor) {
-        super(Material.ROCK, brickColor.getName() + "_stone_brick", new BlockProperties(10.0f, 5.0f, ToolType.PICKAXE, 0));
+        super(Material.ROCK, brickColor.getName() + "_stone_brick", new BlockProperties(10.0f, 5.0f, blocks.stone_bricks.props));
         this.color = brickColor.getMapColor();
         this.name = brickColor.getName() + "_stone_brick";
     }