+++ src/main/java/com/sofodev/armorplus/blocks/normal/BlockCompressedObsidian.java
@@ -6,7 +6,6 @@ package com.sofodev.armorplus.blocks.normal;
 
 import com.sofodev.armorplus.blocks.BlockProperties;
 import com.sofodev.armorplus.blocks.base.BlockBase;
-import com.sofodev.armorplus.blocks.base.ToolType;
 import com.sofodev.armorplus.iface.IModdedBlock;
 import net.minecraft.block.material.MapColor;
 import net.minecraft.block.material.Material;
@@ -23,6 +22,8 @@ import net.minecraft.world.World;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 
+import static com.sofodev.armorplus.config.ModConfig.RegistryConfig.blocks;
+
 /**
  * @author Sokratis Fotkatzikis
  */
@@ -31,7 +32,7 @@ public class BlockCompressedObsidian extends BlockBase implements IModdedBlock {
     public static final PropertyDirection FACING = PropertyDirection.create("facing", EnumFacing.Plane.HORIZONTAL);
 
     public BlockCompressedObsidian() {
-        super(Material.ROCK, "compressed_obsidian", new BlockProperties(2000.0f, 50.0f, ToolType.PICKAXE, 3));
+        super(Material.ROCK, "compressed_obsidian", new BlockProperties(2000.0f, 50.0f, blocks.block_compressed_obsidian.props));
         this.setDefaultState(this.blockState.getBaseState().withProperty(FACING, EnumFacing.NORTH));
     }
 
@@ -73,6 +74,7 @@ public class BlockCompressedObsidian extends BlockBase implements IModdedBlock {
         return new BlockStateContainer(this, FACING);
     }
 
+    @SuppressWarnings("deprecation")
     @Override
     public Material getMaterial(IBlockState state) {
         return super.getMaterial(state);